quciker installation, merge installation and apply configs

+++ haproxy/install.sh
@@ -1,17 +1,19 @@
-systemctl kill hiddify-sniproxy > /dev/null 2>&1
-systemctl stop hiddify-sniproxy > /dev/null 2>&1
-systemctl disable hiddify-sniproxy > /dev/null 2>&1
-
-systemctl kill haproxy > /dev/null 2>&1
-systemctl stop haproxy > /dev/null 2>&1
-systemctl disable haproxy > /dev/null 2>&1
-# pkill -9 haproxy
-pkill -9 sniproxy > /dev/null 2>&1
-if ! command -v haproxy ;then
-    add-apt-repository ppa:vbernat/haproxy-2.8 -y
-    apt update
-    apt install haproxy=2.8.\* -y
+source ../common/utils.sh
+if is_installed sniproxy; then
+    # systemctl kill hiddify-sniproxy > /dev/null 2>&1
+    systemctl stop hiddify-sniproxy >/dev/null 2>&1
+    systemctl disable hiddify-sniproxy >/dev/null 2>&1
+    pkill -9 sniproxy >/dev/null 2>&1
 fi
+
+if ! is_installed haproxy; then
+    add-apt-repository ppa:vbernat/haproxy-2.8 -y
+    install_package haproxy
+fi
+systemctl kill haproxy >/dev/null 2>&1
+systemctl stop haproxy >/dev/null 2>&1
+systemctl disable haproxy >/dev/null 2>&1
+
 ln -sf $(pwd)/hiddify-haproxy.service /etc/systemd/system/hiddify-haproxy.service
 systemctl enable hiddify-haproxy.service
